BAP-8775: Autoresponse jobs are created without autoresponse rules. Fix migration from 1.5 to 1.8

+++ src/Oro/Bundle/EmailBundle/Migrations/Schema/v1_17/OroEmailBundle.php
@@ -15,17 +15,15 @@ class OroEmailBundle implements Migration
      */
     public function up(Schema $schema, QueryBag $queries)
     {
-        $queries->addQuery(new ParametrizedSqlMigrationQuery(
-            'DELETE FROM oro_process_definition WHERE name = :name',
-            [
-                'name' => 'email_auto_response',
-            ]
-        ));
-        $queries->addQuery(new ParametrizedSqlMigrationQuery(
-            'DELETE FROM oro_process_trigger WHERE definition_name = :name',
-            [
-                'name' => 'email_auto_response',
-            ]
-        ));
+        if ($schema->hasTable('oro_process_definition')) {
+            $queries->addQuery(new ParametrizedSqlMigrationQuery(
+                'DELETE FROM oro_process_definition WHERE name = :name',
+                ['name' => 'email_auto_response']
+            ));
+            $queries->addQuery(new ParametrizedSqlMigrationQuery(
+                'DELETE FROM oro_process_trigger WHERE definition_name = :name',
+                ['name' => 'email_auto_response']
+            ));
+        }
     }
 }
